+++ CHANGES.md
@@ -1,11 +1,12 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
-Changes in 12.8
----------------
+Changes in 12.7.2
+-----------------
 - Fix Checkbox input, use `_makeCheckboxInput()` function in Schools.php
 - Fix CSS responsive header, use .rseparator class in Grades.php & AnomalousGrades.php
 - Fix SQL error invalid input syntax for type date: "" in AssignOtherInfo.php
+- Fix Final Grade when assignments are weighted in StudentGrades.php
 
 Changes 12.7.1
 --------------